+++ scripts/applications/rollResolverSingle.js
@@ -100,12 +100,14 @@ export class CPRSingleRollResolver extends HandlebarsApplicationMixin(Applicatio
     }
     async close(options={}) {
         // eslint-disable-next-line no-undef
-        if ( this.rendered ) await this.constructor._fulfillRoll.call(this, null, null, new FormDataExtended(this.element));
+        if ( this.rendered ) await this.constructor._fulfillRoll.call(this, null, null, new foundry.applications.ux.FormDataExtended(this.element));
         Roll.defaultImplementation.RESOLVERS.delete(this.roll);
         this.#resolve?.();
         return super.close(options);
     }
     async _prepareContext(_options) {
+        const diceOnly = genericUtils.getCPRSetting('manualRollsInputDiceOnly');
+        const placeholderKey = diceOnly ? 'CHRISPREMADES.Generic.DiceOnly' : 'CHRISPREMADES.Generic.Total';
         const context = {
             formula: this.roll.formula,
             groups: [{
@@ -129,7 +131,8 @@ export class CPRSingleRollResolver extends HandlebarsApplicationMixin(Applicatio
                     if (cur instanceof CONFIG.Dice.termTypes.NumericTerm) acc += cur.number;
                 }, 0)
             },
-            buttons: [{type: 'submit', label: 'CHRISPREMADES.Generic.Submit', name: 'confirm', icon: 'fa-solid fa-check'}]
+            buttons: [{type: 'submit', label: 'CHRISPREMADES.Generic.Submit', name: 'confirm', icon: 'fa-solid fa-check'}],
+            placeholder: genericUtils.translate(placeholderKey) //placeholder text depending on 'manualRollsInputDiceOnly' setting
         };
         context.options.content = (!context.options.name || !context.options.type) ? context.formula : context.options.name + ' - ' + context.options.type;
         if (this.roll.options?.rollType?.toLowerCase()?.includes('attack')) context.quickButtons = [
@@ -196,11 +199,13 @@ export class CPRSingleRollResolver extends HandlebarsApplicationMixin(Applicatio
                         dice.max += (die.faces * dieAmount);
                         for (let i = 0; i < dieAmount; i++) {
                             dice.terms.push(die.faces);
-                        }   
+                        }
                     } else if (die instanceof CONFIG.Dice.termTypes.OperatorTerm) {
                         dice.multiplier = die.operator === '-' ? -1 : 1;
                     } else if (die instanceof CONFIG.Dice.termTypes.NumericTerm) {
-                        total -= (die.number * dice.multiplier);
+                        if (!genericUtils.getCPRSetting('manualRollsInputDiceOnly')) {
+                            total -= (die.number * dice.multiplier);
+                        }
                     }
                     return dice;
                 }, {terms: [], max: 0, multiplier: 1}));
@@ -211,7 +216,7 @@ export class CPRSingleRollResolver extends HandlebarsApplicationMixin(Applicatio
                 } else results = (dice.terms.reduce((results, number) => {
                     results.diceLeft -= 1;
                     if (number + results.diceLeft <= total) {
-                        let value = ((number === this.roll?.options?.critical) && (total != dice.max)) ? number - 1 : number; 
+                        let value = ((number === this.roll?.options?.critical) && (total != dice.max)) ? number - 1 : number;
                         results.diceArray.push({faces: number, result: value});
                         total -= value;
                     } else if (1 + results.diceLeft >= total) {