+++ scripts/macros.js
@@ -302,5 +302,11 @@ export {assassinate} from './macros/2024/classFeatures/rogue/assassin/assassinat
 export {infiltrationExpertise} from './macros/2024/classFeatures/rogue/assassin/infiltrationExpertise.js';
 export {envenomWeapons, envenomPoison} from './macros/2024/classFeatures/rogue/assassin/envenomWeapons.js';
 export {deathStrike} from './macros/2024/classFeatures/rogue/assassin/deathStrike.js';
+// Soulknife
+export {psionicPower} from './macros/2024/classFeatures/rogue/soulknife/psionicPower.js';
+export {psychicBlades} from './macros/2024/classFeatures/rogue/soulknife/psychicBlades.js';
+export {soulBlades} from './macros/2024/classFeatures/rogue/soulknife/soulBlades.js';
+export {psychicVeil} from './macros/2024/classFeatures/rogue/soulknife/psychicVeil.js';
+export {rendMind} from './macros/2024/classFeatures/rogue/soulknife/rendMind.js';
 //Multiple
 export {evasion} from './macros/2024/classFeatures/multiple/evasion.js';